+++ cpp/SearchInsertPosition.cpp
@@ -5,10 +5,10 @@ using namespace std;
 class Solution {
  public:
   int searchInsert(vector<int> &nums, int target) {
-    for (int i = 0; i < (int)nums.size(); i++) {
+    for (unsigned long i = 0; i < nums.size(); i++) {
       if (nums[i] == target || nums[i] > target) return i;
     }
-    return nums.size();
+    return (int)nums.size();
   }
 
   void output(vector<int> &nums, int target) {
